part1_mock_interview/requirements.txt
livekit-agents[deepgram,openai,cartesia,silero,turn-detector]>=1.0
python-dotenv
